refactor: drop every #[allow] via real types, resolve stale TEMPORARY labels
Rule 1 of this project is no warnings and no `#[allow]`. Seven had
accumulated. Each is now removed by extracting the type the signature was
missing rather than by silencing the lint:

  - grain_contact_law: `RollingParams`, `ContactKinematics`
  - cfl: `SubstepScene<'a>`, `SubstepBounds`
  - surface_reconstruction: `PhasePipelineBuffers<'a>` (10 wgpu buffers)
  - registry: two stale allows simply deleted

Also resolves labels whose investigations have concluded, so the words
mean what they say:

  - `Particles::contact_count` was "TEMP DIAGNOSTIC, remove once the
    investigation concludes". It is the coordination number, a standard
    granular-physics measure, and four real tests consume it. Kept as
    permanent API and documented as such.
  - mixture/pressure's env-gated probe is a permanent opt-in debugging
    facility, not a pending cleanup.
  - fluid.rs's "TEMPORARY restoration" is proven necessary and covered by
    its own tests.

Plus real clippy debt found by actually running `cargo clippy -D warnings`.

+++ tests/self_gravitating_body.rs
@@ -154,7 +154,7 @@ fn radius_of_gyration(sim: &Simulation) -> f32 {
 #[test]
 fn regolith_body_stays_gravitationally_bound_under_self_gravity() {
     let mut solver = make_body();
-    let g_grid = (6.674e-11 / (DX_METERS * DX_METERS * DX_METERS)) as f64;
+    let g_grid = 6.674e-11 / (DX_METERS * DX_METERS * DX_METERS);
     let n = solver.particles().len();
 
     let total_mass_kg: f64 = (0..n).map(|i| solver.particles().mass[i] as f64).sum();
